+++ changes.md
@@ -2,6 +2,10 @@
 
 Here **[Breaking]** means the change that may break valid user code, **[Fix]** - a bug fix.
 
+## 0.1.6 - 2023-12-19
+* **[New]** Added Python examples in python/examples
+* **[Fix]** Python: `get_solution_dict` and `getSolutionDict`
+
 ## 0.1.5 - 2023-11-20
 
 * **[Fix]** Gurobi: getting solution vector and objective bounds in callbacks at MIPSOL and MIPNODE